Added Docs folder for .md and Obsidian files
-Removed /Docs from .gitignore
-Added sample Obsidian workspace files
-Added generic documents to /Docs/Engine

These docs may go into the cloud or a project management system, but that is TBD. If the docs remain simple, may just keep in the repo.

In the future I will place binaries, assets (art, sound, etc), and docs to go elsewhere. Repos are for code.

+++ Engine/Core/Ids.h
@@ -16,7 +16,7 @@ namespace rogue
 
 	struct ModuleId
 	{
-		uint16_t value = 0;
+		uint8_t value = 0;
 		constexpr bool IsValid() const
 		{
 			return value != 0;
